Audit - L08 combined integration-lib and program
