--- solana/integration-lib/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "solana-gateway"
-version = "0.2.0"
-description = "Solana on-chain identity gateway"
-repository = "https://github.com/identity-com/on-chain-identity-gateway"
-authors = ["Identity.com <daniel@identity.com>"]
-license = "MIT"
-edition = "2018"
-keywords = ["solana", "identity",  "gateway", "civic"]
-homepage = "https://www.identity.com"
-
-
-[features]
-program = []
-client = []
-test = []
-default = ["program"]
-no-entrypoint = []
-
-[dependencies]
-# This must stay on 0.9.* as solana-program is using 0.9.1
-borsh = "0.9.1"
-solana-program = "1.15.2"
-num-derive = "0.3"
-num-traits = "0.2.12"
-thiserror = "1.0.20"
-bitflags = "2.1.0"
-
-[dev-dependencies]
-solana-sdk = "1.15.2"
-solana-gateway-program = { path = "../program" }
