+++ solana/program/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 
 [features]
 no-entrypoint = []
-test-bpf = []
 test-sbf = []
 
 [dependencies]
@@ -18,9 +17,7 @@ borsh = "0.9.1"
 num-derive = "0.3"
 num-traits = "0.2"
 solana-program = "1.15.2"
-solana-gateway = { path = "../integration-lib" }
 thiserror = "1.0"
-zeroize = "=1.3.0"
 
 [dev-dependencies]
 solana-program-test = "1.15.2"
