+++ solana/program/src/entrypoint.rs
@@ -1,13 +1,16 @@
 //! Program entrypoint
-use solana_program::{
-    account_info::AccountInfo, entrypoint, entrypoint::ProgramResult, pubkey::Pubkey,
-};
+#[cfg(not(feature = "no-entrypoint"))]
+pub mod entrypoint {
+    use solana_program::{
+        account_info::AccountInfo, entrypoint, entrypoint::ProgramResult, pubkey::Pubkey,
+    };
 
-entrypoint!(process_instruction);
-pub fn process_instruction(
-    program_id: &Pubkey,
-    accounts: &[AccountInfo],
-    instruction_data: &[u8],
-) -> ProgramResult {
-    crate::processor::process_instruction(program_id, accounts, instruction_data)
+    entrypoint!(process_instruction);
+    pub fn process_instruction(
+        program_id: &Pubkey,
+        accounts: &[AccountInfo],
+        instruction_data: &[u8],
+    ) -> ProgramResult {
+        crate::processor::process_instruction(program_id, accounts, instruction_data)
+    }
 }
