+++ solana/program/src/test_utils.rs
@@ -8,6 +8,7 @@ pub mod test_utils_stubs {
     use solana_program::program_stubs;
     use std::sync::Once;
     use std::time::{SystemTime, UNIX_EPOCH};
+    use crate::processor::process_instruction;
 
     static INIT_TESTS: Once = Once::new();
 
@@ -51,7 +52,7 @@ pub mod test_utils_stubs {
             _signers_seeds: &[&[&[u8]]],
         ) -> ProgramResult {
             if instruction.program_id == Gateway::program_id() {
-                solana_gateway_program::processor::process_instruction(
+                process_instruction(
                     &instruction.program_id,
                     account_infos,
                     &instruction.data,
